+++ src/main/java/com/c2psi/businessmanagement/models/PackingDetails.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -18,9 +19,9 @@ import javax.validation.constraints.PositiveOrZero;
         columnNames = {"packaging_id", "loading_id"})})
 public class PackingDetails extends AbstractEntity{
 
-    Integer pdNumberofpackageused;
+    BigDecimal pdNumberofpackageused;
 
-    Integer pdNumberofpackagereturn;
+    BigDecimal pdNumberofpackagereturn;
 
     /******************************
      * Relation between entities  *
