+++ src/main/java/com/c2psi/businessmanagement/models/UnitConversion.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.PositiveOrZero;
+import java.math.BigDecimal;
 
 @Data
 @NoArgsConstructor
@@ -19,7 +20,7 @@ import javax.validation.constraints.PositiveOrZero;
                 columnNames = {"unit_source_id", "unit_destination_id"})})
 public class UnitConversion extends AbstractEntity {
 
-    Double conversionFactor;
+    BigDecimal conversionFactor;
     /******************************
      * Relation between entities  *
      * ****************************/
